+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 Documentation for MIVisionX is available at
 [https://rocm.docs.amd.com/projects/MIVisionX/en/latest/doxygen/html/index.html](https://rocm.docs.amd.com/projects/MIVisionX/en/latest/doxygen/html/index.html)
 
-## MIVisionX 3.0.0 (unreleased)
+## MIVisionX 3.0.0
 
 ### Additions
 
@@ -32,6 +32,7 @@ Documentation for MIVisionX is available at
 ### Fixes
 
 * Dependencies
+* Setup bugfix
 
 ### Tested configurations
 
@@ -48,11 +49,12 @@ Documentation for MIVisionX is available at
 * OpenCV - [4.6.0](https://github.com/opencv/opencv/releases/tag/4.6.0)
 * FFMPEG - [n4.4.2](https://github.com/FFmpeg/FFmpeg/releases/tag/n4.4.2)
 * Dependencies for all the above packages
-* MIVisionX Setup Script - `V3.5.0`
+* MIVisionX Setup Script - `V3.6.0`
 
 ### Known issues
 
-* MIVisionX package install requires manual prerequisites installation
+* Package install requires `OpenCV V4.6` manual install
+* CentOS/RedHat/SLES requires `FFMPEG Dev` package manual install
 
 ## MIVisionX 2.5.0